attach form submission button to sibling form and move event to onSubmit
